Issue 395 COMMITTING NEW FILES WARNING: Large refactoring: made oss-commons a superset of oss-commons + commercial 'commons' such that we can do away with commercial 'commons'.  In the process, I took mostly the latest code from 'commons'.  I then went through each of the packages in oss-commmons and compared them with the base revision to make sure that critical packages like config and mysql did not regress.  I believe I've got it right.
This commit also causes the build to no longer build the 'commercial' commons. For this reason, it builds the open source replicator first - we will be using the oss-commons from there - and then builds the commercial code which references the oss-commons.

NOTE: I believe that oss-commons should become a project, now, in it's own right and be built independently.



git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@1028 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928