Update Issue 950 Update Issue 951
Update tungsten_provision_slave to work on replicators configured with the 'direct' topology. It will automatically detect this and use the master as the source of information. The replication position will be set to the correct location before the replication service is brought back online.

Update tungsten_set_position and tungsten_get_position to work with Redshift.


git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@2460 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928